♻️ (dashboard): Move tick formatter to actionVolumeUtils
Relocate makeTickFormatter (with its out-of-range guard) out of the panel
component into the utils module — mirrors transformSeries and keeps the
panel file free of react-refresh non-component exports.

Refs AAASM-4155
